Save MRN No in time inmilliseconds

+++ app/src/main/java/org/ei/telemedicine/view/contract/ECSmartRegisterBaseClient.java
@@ -11,5 +11,5 @@ public interface ECSmartRegisterBaseClient {
         }
     };
 
-    public Integer ecNumber();
+    public Long ecNumber();
 }
